Refactor header and login page style
