+++ app/scripts/components/landing-page/main.jsx
@@ -3,33 +3,27 @@
 
 	var React = require('react'),
 		RouteHandler = require('react-router').RouteHandler,
-		Header = require('./header.jsx'),
-		UserStore = require('../../stores/userStore'),
-		UserActions = require('../../actions/userActions');
+		Header = require('./header.jsx');
 
-		var Main = React.createClass({
+		module.exports = React.createClass({
 			getInitialState: function() {
 				return {
 					user: {}
 				};
 			},
 
-			componentDidMount: function() {
-				UserActions.session();
-				UserStore.addChangeListener(this.getSession);
-			},
-
-			getSession: function () {
-				var data = UserStore.getData();
-				if(data && !data.error) {
-					this.setState({user: data});
+			setUser: function(user) {
+				if(user) {
+					this.setState({
+						user: user
+					});
 				}
 			},
 
 			render: function() {
 				return (
 					<div>
-						<Header user={this.state.user} />
+						<Header setUser={this.setUser} user={this.state.user} />
 						<div className="handler">
 							<RouteHandler />
 						</div>
@@ -37,6 +31,4 @@
 				);
 			}
 		});
-
-		module.exports = Main;
 })();
